Fix `exclude` option (was negated) and fix callback invocation
The plugin callback was invoked once per asset. Instead we need to invoke it just once, so use `Promise.all` to wait for all of the assets to be compiled before invoking the callback. Eventually we should subscribe to events for individual assets so that we don't re-emit all of the asset modules on every build when using `--watch`.